+++ checker/src/main/java/org/checkerframework/checker/signedness/SignednessVisitor.java
@@ -6,11 +6,8 @@ import com.sun.source.tree.ExpressionTree;
 import com.sun.source.tree.MethodInvocationTree;
 import com.sun.source.tree.MethodTree;
 import com.sun.source.tree.Tree;
-import java.util.Collections;
-import java.util.Set;
 import javax.lang.model.element.AnnotationMirror;
 import javax.lang.model.element.ExecutableElement;
-import javax.lang.model.type.TypeKind;
 import org.checkerframework.checker.interning.InterningVisitor;
 import org.checkerframework.checker.interning.qual.EqualsMethod;
 import org.checkerframework.checker.signedness.qual.PolySigned;
@@ -20,6 +17,7 @@ import org.checkerframework.common.basetype.BaseTypeChecker;
 import org.checkerframework.common.basetype.BaseTypeVisitor;
 import org.checkerframework.framework.type.AnnotatedTypeMirror;
 import org.checkerframework.framework.type.AnnotatedTypeMirror.AnnotatedExecutableType;
+import org.checkerframework.javacutil.AnnotationMirrorSet;
 import org.checkerframework.javacutil.BugInCF;
 import org.checkerframework.javacutil.Pair;
 import org.checkerframework.javacutil.TreeUtils;
@@ -140,13 +138,10 @@ public class SignednessVisitor extends BaseTypeVisitor<SignednessAnnotatedTypeFa
           AnnotationMirror leftAnno = leftOpType.getEffectiveAnnotations().iterator().next();
           AnnotationMirror rightAnno = rightOpType.getEffectiveAnnotations().iterator().next();
 
-          if (leftOpType.getKind() != TypeKind.CHAR
-              && !TypesUtils.isDeclaredOfName(leftOpType.getUnderlyingType(), "java.lang.Character")
+          if (!TypesUtils.isCharOrCharacter(leftOpType.getUnderlyingType())
               && !atypeFactory.getQualifierHierarchy().isSubtype(leftAnno, atypeFactory.SIGNED)) {
             checker.reportError(leftOp, "unsigned.concat");
-          } else if (rightOpType.getKind() != TypeKind.CHAR
-              && !TypesUtils.isDeclaredOfName(
-                  rightOpType.getUnderlyingType(), "java.lang.Character")
+          } else if (!TypesUtils.isCharOrCharacter(rightOpType.getUnderlyingType())
               && !atypeFactory.getQualifierHierarchy().isSubtype(rightAnno, atypeFactory.SIGNED)) {
             checker.reportError(rightOp, "unsigned.concat");
           }
@@ -309,13 +304,12 @@ public class SignednessVisitor extends BaseTypeVisitor<SignednessAnnotatedTypeFa
 
       case PLUS_ASSIGNMENT:
         if (TreeUtils.isStringCompoundConcatenation(tree)) {
-          if (exprType.getKind() != TypeKind.CHAR
-              && !TypesUtils.isDeclaredOfName(
-                  exprType.getUnderlyingType(), "java.lang.Character")) {
-            AnnotationMirror anno = exprType.getEffectiveAnnotations().iterator().next();
-            if (!atypeFactory.getQualifierHierarchy().isSubtype(anno, atypeFactory.SIGNED)) {
-              checker.reportError(tree.getExpression(), "unsigned.concat");
-            }
+          if (TypesUtils.isCharOrCharacter(exprType.getUnderlyingType())) {
+            break;
+          }
+          AnnotationMirror anno = exprType.getEffectiveAnnotations().iterator().next();
+          if (!atypeFactory.getQualifierHierarchy().isSubtype(anno, atypeFactory.SIGNED)) {
+            checker.reportError(tree.getExpression(), "unsigned.concat");
           }
           break;
         }
@@ -352,8 +346,8 @@ public class SignednessVisitor extends BaseTypeVisitor<SignednessAnnotatedTypeFa
   }
 
   @Override
-  protected Set<? extends AnnotationMirror> getExceptionParameterLowerBoundAnnotations() {
-    return Collections.singleton(atypeFactory.SIGNED);
+  protected AnnotationMirrorSet getExceptionParameterLowerBoundAnnotations() {
+    return new AnnotationMirrorSet(atypeFactory.SIGNED);
   }
 
   @Override