+++ framework/src/main/java/org/checkerframework/common/value/ValueCheckerUtils.java
@@ -17,7 +17,6 @@ import org.checkerframework.framework.type.AnnotatedTypeFactory;
 import org.checkerframework.framework.type.AnnotatedTypeMirror;
 import org.checkerframework.framework.type.GenericAnnotatedTypeFactory;
 import org.checkerframework.javacutil.AnnotationUtils;
-import org.checkerframework.javacutil.SystemUtil;
 import org.checkerframework.javacutil.TypeSystemError;
 import org.checkerframework.javacutil.TypesUtils;
 import org.plumelib.util.CollectionsPlume;
@@ -280,7 +279,7 @@ public class ValueCheckerUtils {
    */
   public static List<Integer> getLengthsForStringValues(List<String> values) {
     List<Integer> lengths = CollectionsPlume.mapList(String::length, values);
-    return SystemUtil.withoutDuplicatesSorted(lengths);
+    return CollectionsPlume.withoutDuplicatesSorted(lengths);
   }
 
   /**