+++ framework/src/main/java/org/checkerframework/framework/type/AnnotatedTypeFactory.java
@@ -73,6 +73,7 @@ import javax.lang.model.util.Types;
 import javax.tools.Diagnostic;
 import org.checkerframework.afu.scenelib.el.AMethod;
 import org.checkerframework.afu.scenelib.el.ATypeElement;
+import org.checkerframework.checker.formatter.qual.FormatMethod;
 import org.checkerframework.checker.interning.qual.FindDistinct;
 import org.checkerframework.checker.interning.qual.InternedDistinct;
 import org.checkerframework.checker.nullness.qual.MonotonicNonNull;
@@ -141,6 +142,7 @@ import org.checkerframework.javacutil.trees.DetachedVarSymbol;
 import org.plumelib.util.CollectionsPlume;
 import org.plumelib.util.ImmutableTypes;
 import org.plumelib.util.StringsPlume;
+import org.plumelib.util.SystemPlume;
 
 /**
  * The methods of this class take an element or AST node, and return the annotated type as an {@link
@@ -174,6 +176,9 @@ import org.plumelib.util.StringsPlume;
  * @checker_framework.manual #creating-a-checker How to write a checker plug-in
  */
 public class AnnotatedTypeFactory implements AnnotationProvider {
+
+  /** Whether to output verbose, low-level debugging messages about {@link #getAnnotatedType}. */
+  private static final boolean debugGat = false;
 
   /** Whether to print verbose debugging messages about stub files. */
   private final boolean debugStubParser;
@@ -1308,6 +1313,7 @@ public class AnnotatedTypeFactory implements AnnotationProvider {
    * @return the annotated type of {@code tree}
    */
   public AnnotatedTypeMirror getAnnotatedType(Tree tree) {
+    logGat("getAnnotatedType(%s)%n", TreeUtils.toStringTruncated(tree, 60));
 
     if (tree == null) {
       throw new BugInCF("AnnotatedTypeFactory.getAnnotatedType: null tree");
@@ -1324,16 +1330,25 @@ public class AnnotatedTypeFactory implements AnnotationProvider {
     } else if (TreeUtils.isExpressionTree(tree)) {
       tree = TreeUtils.withoutParens((ExpressionTree) tree);
       type = fromExpression((ExpressionTree) tree);
+      logGat(
+          "getAnnotatedType(%s): fromExpression=>%s%n",
+          TreeUtils.toStringTruncated(tree, 60), type);
     } else {
       throw new BugInCF(
           "AnnotatedTypeFactory.getAnnotatedType: query of annotated type for tree "
               + tree.getKind());
     }
 
+    logGat(
+        "getAnnotatedType(%s): before addComputedTypeAnnotations, type=%s%n",
+        TreeUtils.toStringTruncated(tree, 60), type);
     addComputedTypeAnnotations(tree, type);
     if (tree.getKind() == Kind.TYPE_CAST) {
       type = applyCaptureConversion(type);
     }
+    logGat(
+        "getAnnotatedType(%s): after addComputedTypeAnnotations, type=%s%n",
+        TreeUtils.toStringTruncated(tree, 60), type);
 
     if (TreeUtils.isClassTree(tree) || tree.getKind() == Tree.Kind.METHOD) {
       // Don't cache VARIABLE
@@ -1420,7 +1435,7 @@ public class AnnotatedTypeFactory implements AnnotationProvider {
    * @return the set of qualifiers that are the upper bound for a type use if no other bound is
    *     specified for the type
    */
-  protected Set<? extends AnnotationMirror> getDefaultTypeDeclarationBounds() {
+  protected AnnotationMirrorSet getDefaultTypeDeclarationBounds() {
     return qualHierarchy.getTopAnnotations();
   }
 
@@ -1641,7 +1656,9 @@ public class AnnotatedTypeFactory implements AnnotationProvider {
    * @see TypeFromExpressionVisitor
    */
   private AnnotatedTypeMirror fromExpression(ExpressionTree tree) {
+    logGat("fromExpression(%s)%n", tree);
     if (shouldCache && fromExpressionTreeCache.containsKey(tree)) {
+      logGat("fromExpression(%s) => [cached] %s%n", tree, fromExpressionTreeCache.get(tree));
       return fromExpressionTreeCache.get(tree).deepCopy();
     }
 
@@ -1655,6 +1672,7 @@ public class AnnotatedTypeFactory implements AnnotationProvider {
       // cached during dataflow analysis. See Issue #602.
       fromExpressionTreeCache.put(tree, result.deepCopy());
     }
+    logGat("fromExpression(%s) => %s%n", tree, result);
     return result;
   }
 
@@ -5839,5 +5857,19 @@ public class AnnotatedTypeFactory implements AnnotationProvider {
       }
     }
     return false;
+  }
+
+  /**
+   * Output a message about {@link #getAnnotatedType}, if logging is on.
+   *
+   * @param format a format string
+   * @param args arguments to the format string
+   */
+  @FormatMethod
+  public void logGat(String format, Object... args) {
+    if (debugGat) {
+      SystemPlume.sleep(1); // logging can interleave with typechecker output
+      System.out.printf(format, args);
+    }
   }
 }