+++ framework/src/main/java/org/checkerframework/framework/type/GenericAnnotatedTypeFactory.java
@@ -145,6 +145,12 @@ public abstract class GenericAnnotatedTypeFactory<
         FlowAnalysis extends CFAbstractAnalysis<Value, Store, TransferFunction>>
     extends AnnotatedTypeFactory {
 
+  /**
+   * Whether to output verbose, low-level debugging messages. Also see {@code TreeAnnotator.debug}
+   * and {@link AnnotatedTypeFactory#debugGat}.
+   */
+  private static final boolean debug = false;
+
   /** To cache the supported monotonic type qualifiers. */
   private @MonotonicNonNull Set<Class<? extends Annotation>> supportedMonotonicQuals;
 
@@ -625,14 +631,17 @@ public abstract class GenericAnnotatedTypeFactory<
   }
 
   /**
-   * Returns the appropriate transfer function that is used for the org.checkerframework.dataflow
-   * analysis.
+   * Returns the appropriate transfer function that is used for the given
+   * org.checkerframework.dataflow analysis.
    *
    * <p>This implementation uses the checker naming convention to create the appropriate transfer
    * function. If no transfer function is found, it returns an instance of {@link CFTransfer}.
    *
    * <p>Subclasses have to override this method to create the appropriate transfer function if they
    * do not follow the checker naming convention.
+   *
+   * @param analysis a dataflow analysis
+   * @return a new transfer function
    */
   // A more precise type for the parameter would be FlowAnalysis, which
   // is the type parameter bounded by the current parameter type CFAbstractAnalysis<Value, Store,
@@ -1844,14 +1853,18 @@ public abstract class GenericAnnotatedTypeFactory<
     applyQualifierParameterDefaults(tree, type);
     log("%s GATF.addComputedTypeAnnotations#3(%s, %s)%n", thisClass, treeString, type);
     treeAnnotator.visit(tree, type);
-    log("%s GATF.addComputedTypeAnnotations#4(%s, %s)%n", thisClass, treeString, type);
+    log(
+        "%s GATF.addComputedTypeAnnotations#4(%s, %s)%n  treeAnnotator=%s%n",
+        thisClass, treeString, type, treeAnnotator);
     if (TreeUtils.isExpressionTree(tree)) {
       // If a tree annotator, did not add a type, add the DefaultForUse default.
       addAnnotationsFromDefaultForType(TreeUtils.elementFromTree(tree), type);
       log("%s GATF.addComputedTypeAnnotations#5(%s, %s)%n", thisClass, treeString, type);
     }
     typeAnnotator.visit(type, null);
-    log("%s GATF.addComputedTypeAnnotations#6(%s, %s)%n", thisClass, treeString, type);
+    log(
+        "%s GATF.addComputedTypeAnnotations#6(%s, %s)%n  typeAnnotator=%s%n",
+        thisClass, treeString, type, typeAnnotator);
     defaults.annotate(tree, type);
     log("%s GATF.addComputedTypeAnnotations#7(%s, %s)%n", thisClass, treeString, type);
 
@@ -2318,10 +2331,6 @@ public abstract class GenericAnnotatedTypeFactory<
     }
   }
 
-  // You can change this temporarily to produce verbose logs.
-  /** Whether to output verbose, low-level debugging messages. */
-  private static final boolean debug = false;
-
   /**
    * Output a message, if logging is on.
    *
@@ -2336,12 +2345,8 @@ public abstract class GenericAnnotatedTypeFactory<
     }
   }
 
-  /**
-   * Cache of types found that are relevantTypes or subclass of supported types. Used so that
-   * isSubtype doesn't need to be called repeatedly on the same types.
-   */
-  private final Map<TypeMirror, Boolean> allFoundRelevantTypes =
-      CollectionUtils.createLRUCache(300);
+  /** For each type, whether it is relevant. A cache to avoid repeated re-computation. */
+  private final Map<TypeMirror, Boolean> isRelevantCache = CollectionUtils.createLRUCache(300);
 
   /**
    * Returns true if users can write type annotations from this type system on the given Java type.
@@ -2351,13 +2356,23 @@ public abstract class GenericAnnotatedTypeFactory<
    */
   public boolean isRelevant(TypeMirror tm) {
     tm = types.erasure(tm);
-    Boolean cachedResult = allFoundRelevantTypes.get(tm);
+    Boolean cachedResult = isRelevantCache.get(tm);
     if (cachedResult != null) {
       return cachedResult;
     }
     boolean result = isRelevantHelper(tm);
-    allFoundRelevantTypes.put(tm, result);
+    isRelevantCache.put(tm, result);
     return result;
+  }
+
+  /**
+   * Returns true if users can write type annotations from this type system on the given type.
+   *
+   * @param tm a type
+   * @return true if users can write type annotations from this type system on the given type
+   */
+  public final boolean isRelevant(AnnotatedTypeMirror tm) {
+    return isRelevant(tm.getUnderlyingType());
   }
 
   /**
@@ -2369,7 +2384,7 @@ public abstract class GenericAnnotatedTypeFactory<
    */
   private boolean isRelevantHelper(TypeMirror tm) {
 
-    if (relevantJavaTypes.contains(tm)) {
+    if (relevantJavaTypes == null || relevantJavaTypes.contains(tm)) {
       return true;
     }
 
@@ -2409,6 +2424,35 @@ public abstract class GenericAnnotatedTypeFactory<
 
       case TYPEVAR:
         return isRelevant(((TypeVariable) tm).getUpperBound());
+
+      case NULL:
+        for (TypeMirror relevantJavaType : relevantJavaTypes) {
+          switch (relevantJavaType.getKind()) {
+            case BOOLEAN:
+            case BYTE:
+            case CHAR:
+            case DOUBLE:
+            case FLOAT:
+            case INT:
+            case LONG:
+            case SHORT:
+              continue;
+
+            case ERROR:
+            case NONE:
+            case VOID:
+              continue;
+
+            case MODULE:
+            case PACKAGE:
+              continue;
+
+            case NULL:
+            default:
+              return true;
+          }
+        }
+        return false;
 
       default:
         throw new BugInCF("isRelevantHelper(%s): Unexpected TypeKind %s", tm, tm.getKind());