+++ framework/tests/framework/MissingSymbolCrash.java
@@ -1,4 +1,3 @@
-// :: error: (type.checking.not.run)
 public class MissingSymbolCrash {
   public void test() {
     // :: error: cannot find symbol