+++ framework/tests/purity-suggestions/PuritySuggestionsClass.java
@@ -8,13 +8,14 @@ public class PuritySuggestionsClass {
 
   String f1, f2, f3;
   String[] a;
+  static String staticString;
 
   // class with a (potentially) non-pure constructor
   private static class NonPureClass {
     String t;
 
     public NonPureClass() {
-      t = "";
+      staticString = "";
     }
   }
 
@@ -22,6 +23,16 @@ public class PuritySuggestionsClass {
   private static class PureClass {
     // :: warning: (purity.more.sideeffectfree)
     public PureClass() {}
+  }
+
+  // class with a pure constructor
+  private static class PureClass2 {
+    String t;
+
+    // :: warning: (purity.more.sideeffectfree)
+    public PureClass2() {
+      t = "";
+    }
   }
 
   // :: warning: (purity.more.sideeffectfree)