+++ javacutil/src/main/java/org/checkerframework/javacutil/TreeUtils.java
@@ -63,6 +63,7 @@ import com.sun.tools.javac.tree.JCTree.JCTypeParameter;
 import com.sun.tools.javac.tree.TreeInfo;
 import com.sun.tools.javac.tree.TreeMaker;
 import com.sun.tools.javac.util.Context;
+import com.sun.tools.javac.util.Position;
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
 import java.util.ArrayList;
@@ -1456,22 +1457,43 @@ public final class TreeUtils {
    * @return true iff if tree is a field access expression (implicit or explicit)
    */
   public static boolean isFieldAccess(Tree tree) {
+    return asFieldAccess(tree) != null;
+  }
+
+  /**
+   * Return the field that {@code tree} is a field access expression for, or null.
+   *
+   * <pre>
+   *   <em>f</em>
+   *   <em>obj</em> . <em>f</em>
+   * </pre>
+   *
+   * This method currently also returns non-null true for class literals and qualified this.
+   *
+   * @param tree a tree that might be a field access
+   * @return the element if tree is a field access expression (implicit or explicit); null otherwise
+   */
+  public static @Nullable VariableElement asFieldAccess(Tree tree) {
     if (tree.getKind() == Tree.Kind.MEMBER_SELECT) {
       // explicit member access (or a class literal or a qualified this)
       MemberSelectTree memberSelect = (MemberSelectTree) tree;
       assert isUseOfElement(memberSelect) : "@AssumeAssertion(nullness): tree kind";
       Element el = TreeUtils.elementFromUse(memberSelect);
-      return el.getKind().isField();
+      if (el.getKind().isField()) {
+        return (VariableElement) el;
+      }
     } else if (tree.getKind() == Tree.Kind.IDENTIFIER) {
       // implicit field access
       IdentifierTree ident = (IdentifierTree) tree;
       assert isUseOfElement(ident) : "@AssumeAssertion(nullness): tree kind";
       Element el = TreeUtils.elementFromUse(ident);
-      return el.getKind().isField()
+      if (el.getKind().isField()
           && !ident.getName().contentEquals("this")
-          && !ident.getName().contentEquals("super");
+          && !ident.getName().contentEquals("super")) {
+        return (VariableElement) el;
+      }
     }
-    return false;
+    return null;
   }
 
   /**
@@ -2401,6 +2423,17 @@ public final class TreeUtils {
       throw new BugInCF(
           "TreeUtils.yieldTreeGetValue: reflection failed for tree: %s", yieldTree, e);
     }
+  }
+
+  /**
+   * Returns true if the {@code variableTree} is declared using the {@code var} Java keyword.
+   *
+   * @param variableTree the variableTree to check
+   * @return true if the variableTree is declared using the {@code var} Java keyword
+   */
+  public static boolean isVariableTreeDeclaredUsingVar(VariableTree variableTree) {
+    JCExpression type = (JCExpression) variableTree.getType();
+    return type != null && type.pos == Position.NOPOS;
   }
 
   /**