+++ javacutil/src/main/java/org/checkerframework/javacutil/TypesUtils.java
@@ -16,6 +16,7 @@ import java.util.Arrays;
 import java.util.Collection;
 import java.util.HashSet;
 import java.util.List;
+import java.util.Locale;
 import java.util.Set;
 import java.util.StringJoiner;
 import javax.annotation.processing.ProcessingEnvironment;
@@ -71,7 +72,7 @@ public final class TypesUtils {
     if (clazz == void.class) {
       return types.getNoType(TypeKind.VOID);
     } else if (clazz.isPrimitive()) {
-      String primitiveName = clazz.getName().toUpperCase();
+      String primitiveName = clazz.getName().toUpperCase(Locale.getDefault());
       TypeKind primitiveKind = TypeKind.valueOf(primitiveName);
       return types.getPrimitiveType(primitiveKind);
     } else if (clazz.isArray()) {
@@ -331,13 +332,25 @@ public final class TypesUtils {
   }
 
   /**
-   * Returns true if the type is either boolean (primitive type) or java.lang.Boolean.
+   * Returns true if the type is either {@code boolean} (primitive type) or {@code
+   * java.lang.Boolean}.
    *
    * @param type the type to test
    * @return true iff type represents a boolean type
    */
   public static boolean isBooleanType(TypeMirror type) {
     return type.getKind() == TypeKind.BOOLEAN || isDeclaredOfName(type, "java.lang.Boolean");
+  }
+
+  /**
+   * Returns true if the type is {@code char} or {@code Character}.
+   *
+   * @param type a type
+   * @return true if the type is {@code char} or {@code Character}
+   */
+  public static boolean isCharOrCharacter(TypeMirror type) {
+    return type.getKind() == TypeKind.CHAR
+        || TypesUtils.isDeclaredOfName(type, "java.lang.Character");
   }
 
   /**
@@ -624,17 +637,6 @@ public final class TypesUtils {
   }
 
   /**
-   * Returns true if {@code type} has an enclosing type.
-   *
-   * @param type type to checker
-   * @return true if {@code type} has an enclosing type
-   */
-  public static boolean hasEnclosingType(TypeMirror type) {
-    Type e = ((Type) type).getEnclosingType();
-    return e.getKind() != TypeKind.NONE;
-  }
-
-  /**
    * Returns whether or not {@code type} is a functional interface type (as defined in JLS 9.8).
    *
    * @param type possible functional interface type
@@ -645,6 +647,17 @@ public final class TypesUtils {
     Context ctx = ((JavacProcessingEnvironment) env).getContext();
     com.sun.tools.javac.code.Types javacTypes = com.sun.tools.javac.code.Types.instance(ctx);
     return javacTypes.isFunctionalInterface((Type) type);
+  }
+
+  /**
+   * Returns true if {@code type} has an enclosing type.
+   *
+   * @param type type to checker
+   * @return true if {@code type} has an enclosing type
+   */
+  public static boolean hasEnclosingType(TypeMirror type) {
+    Type e = ((Type) type).getEnclosingType();
+    return e.getKind() != TypeKind.NONE;
   }
 
   /// Type variables and wildcards